Delete db.sqlite3
IMPORTANT!!! Please see this stackoverflow page for details: https://stackoverflow.com/questions/1274057/how-do-i-make-git-forget-about-a-file-that-was-tracked-but-is-now-in-gitignore

I've set it up so that git stops tracking our database file, so you should expect your LOCAL db.sqlite3 file to be deleted when you pull the latest changes. I've put a copy of it on our shared google drive. If you'd like to save your own copy of the db file, I suggest you do that before you pull changes.